# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Крестики-нолики: серверные запросы",
    "version": "1.0.0",
    "description": "Подробная спецификация серверной части веб-приложения «Крестики-нолики». Сервер поддерживает регистрацию, авторизацию, комнаты, игровые ходы, чат, статистику, достижения, повтор партии, матчмейкинг и административный аудит. Для защищенных запросов используется токен доступа из ответа входа или регистрации.",
    "license": {
      "name": "Учебный курсовой проект"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000",
      "description": "Локальный сервер разработки"
    },
    {
      "url": "http://localhost:5000",
      "description": "Альтернативный локальный адрес"
    }
  ],
...
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": [
          "Авторизация"
        ],
        "summary": "Зарегистрировать пользователя",
        "description": "Создает пользователя, начальную статистику и возвращает токен доступа.",
        "requestBody": {
          "required": true,
          "description": "Данные нового пользователя. Для администратора передайте account_type=admin и admin_code.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              },
              "example": {
                "username": "player_one",
                "password": "secret123",
                "display_name": "Игрок 1",
                "account_type": "client"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Пользователь зарегистрирован",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ok"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "OK"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AuthSession"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ошибка регистрации",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "examples": {
                  "auth_error": {
                    "summary": "Ошибка регистрации",
                    "value": {
                      "ok": false,
                      "error": {
                        "code": "auth_error",
                        "message": "Ошибка регистрации"
                      }
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "examples": {
                  "validation_error": {
                    "summary": "Ошибка валидации",
                    "value": {
                      "ok": false,
                      "error": {
                        "code": "validation_error",
                        "message": "Ошибка валидации"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Авторизация"
        ],
        "summary": "Войти в аккаунт",
        "description": "Проверяет пароль, переводит пользователя online и возвращает новый токен доступа.",
        "requestBody": {
          "required": true,
          "description": "Логин и пароль",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "example": {
                "username": "player_one",
                "password": "secret123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Вход выполнен",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ok"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "OK"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AuthSession"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Неверный логин или пароль",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "examples": {
                  "auth_error": {
                    "summary": "Неверный логин или пароль",
                    "value": {
                      "ok": false,
                      "error": {
                        "code": "auth_error",
                        "message": "Неверный логин или пароль"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "Авторизация"
        ],
        "summary": "Выйти из аккаунта",
        "description": "Отзывает текущий токен доступа.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Успешный ответ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ok"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "OK"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "revoked": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "examples": {
                  "auth_required": {
                    "summary": "Authentication required",
                    "value": {
                      "ok": false,
                      "error": {
                        "code": "auth_required",
                        "message": "Authentication required"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "Авторизация"
        ],
        "summary": "Получить текущего пользователя",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Успешный ответ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ok"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "OK"
                    },
                    "data": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "examples": {
                  "auth_required": {
                    "summary": "Authentication required",
                    "value": {
                      "ok": false,
                      "error": {
                        "code": "auth_required",
                        "message": "Authentication required"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Авторизация"
        ],
        "summary": "Обновить профиль",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Поля профиля для изменения",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateRequest"
              },
              "example": {
                "display_name": "Новый ник",
                "email": "new@example.com"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешный ответ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ok"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "OK"
                    },
                    "data": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "examples": {
                  "auth_required": {
                    "summary": "Authentication required",
                    "value": {
                      "ok": false,
                      "error": {
                        "code": "auth_required",
                        "message": "Authentication required"
                      }
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "examples": {
                  "validation_error": {
                    "summary": "Ошибка валидации",
                    "value": {
                      "ok": false,
                      "error": {
                        "code": "validation_error",
                        "message": "Ошибка валидации"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/users": {
      "get": {
        "tags": [
          "Пользователи"
        ],
        "summary": "Список или поиск пользователей",
        "description": "Без q возвращает последних пользователей, с q ищет по username, display_name и email.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Поисковая строка.",
            "example": "player"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Количество элементов в ответе.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "description": "Смещение для постраничной навигации.",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешный ответ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ok"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "OK"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "examples": {
                  "auth_required": {
                    "summary": "Authentication required",
                    "value": {
                      "ok": false,
                      "error": {
                        "code": "auth_required",
                        "message": "Authentication required"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{user_id}": {
      "get": {
        "tags": [
          "Пользователи"
        ],
        "summary": "Профиль, статистика и достижения пользователя",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "description": "ID пользователя",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Успешный ответ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ok"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "OK"
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserDashboard"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "examples": {
                  "auth_required": {
                    "summary": "Authentication required",
                    "value": {
                      "ok": false,
                      "error": {
                        "code": "auth_required",
                        "message": "Authentication required"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "examples": {
                  "not_found": {
                    "summary": "User not found",
                    "value": {
                      "ok": false,
                      "error": {
                        "code": "not_found",
                        "message": "User not found"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{user_id}/games": {
      "get": {
        "tags": [
          "Пользователи"
        ],
        "summary": "История партий пользователя",
        "description": "Последние партии выбранного пользователя с результатом относительно этого пользователя.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "description": "ID пользователя",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "example": 1
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Количество элементов в ответе.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешный ответ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ok"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "OK"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/RecentGame"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "examples": {
                  "auth_required": {
                    "summary": "Authentication required",
                    "value": {
                      "ok": false,
                      "error": {
                        "code": "auth_required",
                        "message": "Authentication required"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "examples": {
                  "not_found": {
                    "summary": "User not found",
                    "value": {
                      "ok": false,
                      "error": {
                        "code": "not_found",
                        "message": "User not found"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "token",
        "description": "Токен из поля data.token в ответах /api/auth/register и /api/auth/login."
      }
    },
    "schemas": {
...
      "RegisterRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 32,
            "pattern": "^[A-Za-z0-9_]+$",
            "example": "player_one"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 128,
            "format": "password",
            "example": "secret123"
          },
          "display_name": {
            "type": "string",
            "maxLength": 80,
            "example": "Игрок 1"
          },
          "email": {
            "type": "string",
            "format": "email",
            "nullable": true,
            "example": "player@example.com"
          },
          "avatar_url": {
            "type": "string",
            "nullable": true,
            "example": "https://example.com/avatar.png"
          },
          "account_type": {
            "type": "string",
            "enum": [
              "client",
              "admin"
            ],
            "default": "client",
            "example": "client"
          },
          "admin_code": {
            "type": "string",
            "nullable": true,
            "description": "Нужен только для account_type=admin.",
            "example": "admin123"
          }
        }
      },
...
      "AuthSession": {
        "type": "object",
        "required": [
          "token",
          "expires_at",
          "user"
        ],
        "properties": {
          "token": {
            "type": "string",
            "example": "8KpW9kQZbP-token-example"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
...
    }
  },
...
}